### 2021-03-08 #433 @rkrug
ROriginStamp - Interface to OriginStamp API to Obtain Trusted Time Stamps
Submitting Author Name: Rainer M. Krug
Submitting Author Github Handle: @rkrug
Repository: https://github.com/rkrug/ROriginStamp
Version submitted: 0.8.0
Version accepted: TBD

---

```
Package: ROriginStamp
Title: Interface to OriginStamp API to Obtain Trusted Time Stamps
Version: 0.8.0
    person(given = "Rainer M.",
             family = "Krug",
             email = "Rainer.Krug@uzh.ch",
    )
Description: Functions to get and download trusted timestamps and seed files from OriginStamp <https://originstamp.com>.
        curl,
        openssl
    covr,
```

As recommended in the Presubmission inquiry https://github.com/ropensci/software-review/issues/397#issuecomment-703420600

The package automates the process of obtaining trusted timestamps for R objects as well as files.

The package is aimed at a general audience. It can be used to obtain trusted timestamps too prove that any object (e.g.  dataset, graphs, documents of any type, etc) were created at or before the time of the time stamping and not modified afterwards. The hashing of the object and querying this hash at the timestamp provider returns the proof / certificate.

Not to my knowledge.

NA

https://github.com/ropensci/software-review/issues/397#issuecomment-703420600 @maelle

- [X] has [continuous integration](https://ropensci.github.io/dev_guide/ci.html), including reporting of test coverage using services such as Travis CI, Coveralls **and**/~~or~~ CodeCov.
