### 2021-03-11 #433 @noamross
Greetings @rkrug!

### Editor checks:

- [x] **Automated tests:** Package has a testing suite and is tested via a CI service.

---

#### Editor comments

I am unable to run tests successfully locally.  Even when I have set `NOT_CRAN=true` and add `ROriginStamp_api_key` to my `.Renviron`. I think this is likely due to the fact that some of the tests are tied to your own OriginStamp account.  In general we want reviewers/users/contributors to be able to run the test suite.  There are a few options for this, the best of which is generally refactoring the test suite so that any objects that need to be tested against are created as part of the tests using the users' account.  Sometimes testing requires some more elaborate set up like setting up a dummy developers' account.  In such cases where manual steps are required we generally request they be described in a CONTRIBUTING.md file.

I also note that the package vignette isn't visisble at https://rkrug.github.io/ROriginStamp/articles/introduction.html, and I would suggest being explicit in the README that the API key environment variable should be `ROriginStamp_api_key`. I had to find this in the code and we will want this to be simple for the reviewers to get going as easily as possible.

Please make these changes and I can proceed with finding reviewers. Let me know if you have any questions.

Other than that, your package passes `goodpractice::gp()` tests with flying colors, except for a few long code lines, which reviewers can consider when evaluating code readability but requires no immediate action.

```

    R/create_timestamp.R:7:1
    R/create_timestamp.R:8:1
    R/create_timestamp.R:10:1
    R/create_timestamp.R:13:1
    R/create_timestamp.R:14:1
    ... and 58 more lines
```
