+++ README.Rmd
@@ -1,9 +1,6 @@
 ---
 output: github_document
 bibliography: references.bib 
-
-resource_files:
-  - dep_graph.png
 ---
 
 # ROriginStamp - a simple interface to [OriginStamp](https://originstamp.org/)
@@ -85,6 +82,27 @@ devtools::install_github("rkrug/ROriginStamp")
 
 This package makes it easy to obtain and verify a hashed object or file in R, after getting an API key as described above. 
 
+Once you have an API key, you have to make it available to the functions in the package.
+
+This can be done by
+
+1. either set it as an environmental variable, by e.g. using 
+
+```{r, eval = FALSE, include = TRUE}
+Sys.setenv(ROriginStamp_api_key = "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx")
+```     
+or
+
+2. use the `api_key()` function to set the api key by running 
+```{r, eval = FALSE, include = TRUE }
+api_key("xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx")`
+````
+or
+
+3. specify it in each call by using the `key` argument.
+
+
+
 ### Obtain a TTS for an object or file
 
 ```{r eval = FALSE}
@@ -126,4 +144,4 @@ This package aims at minimizing the number of dependencies It has at the moment 
 # } else {
 ```
 
-![Dependency graph](./dep_graph.png)
+![Dependency graph](dep_graph.png)
