### 2021-03-12 f90fb81
Remove docs from gitignore
