### 2021-03-17 17d2310
#433 Fix https://github.com/ropensci/software-review/issues/433#issuecomment-800716196
Fix wrong test snapshot
