+++ tests/testthat/_snaps/006_hash.md
@@ -6,9 +6,7 @@
 
     Code
       tf <- tempfile()
-    Code
       write.csv(data.frame(letters, LETTERS, 1:26), tf)
-    Code
       x <- list(hash = hash(as.hash(
         "2c5d36be542f8f0e7345d77753a5d7ea61a443ba6a9a86bb060332ad56dba38e")),
       r_object = hash(LETTERS), r_object_def = hash.default(LETTERS), file = suppressMessages(
@@ -16,18 +14,14 @@
     Message <simpleMessage>
       
       x is already a hash - returning x unprocessed
-    Message <simpleMessage>
       
       Create sha356 hash from R object x
-    Message <simpleMessage>
       
       Create sha356 hash from R object x
-    Message <simpleMessage>
       
       Create sha356 hash from R object x
     Code
       unlink(tf)
-    Code
       x
     Output
       $hash
