### 2021-03-17 b23fa87
Add ropensci review badge and NEWS.md
