+++ NEWS.md
@@ -1,0 +1,27 @@
+
+ROriginStamp 0.8.0 (2021-03-28)
+=========================
+
+Submission ropensci
+
+ROriginStamp 0.8.x (xxxx-xx-xx)
+=========================
+
+### NEW FEATURES
+
+
+### MINOR IMPROVEMENTS
+
+
+### BUG FIXES
+
+
+### DEPRECATED AND DEFUNCT
+
+
+### DOCUMENTATION FIXES
+
+
+
+### (a special: any heading grouping a large number of changes under one thing)
+
