### 2021-03-18 #433 @cboettig
- [x] **Examples** (that run successfully locally) for all exported functions

No CONTRIBUTING file or guidelines in README

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

Estimated hours spent reviewing: 2 h

---

This is a very nicely implemented package, with a clear coding style, light dependencies, and detailed documentation.

- Function return  display behavior could be cleaned up a bit.  Functions return verbose output that is not easily parsed by most users.  For example, perhaps `create_timestamp` could return a clear message  that timestamp creation was successfully submitted, and indicate how long a user might want to wait before trying to check if the timestamp has registered. For example, the more detailed return information could be placed into an S3 object with a defined print method that displays only key details (hash, time submitted, etc) concisely.

- Minor technical issue, but it's good practice to use put `close()` commands for R connections into `on.exit()` in functions, which allows R to close an open connection if the function terminates prematurely.

- The use of a trademark name has been a concern in the past. From a skim of the legal documentation links provided, it doesn't look like an issue here, but may still be appropriate to check with the owners of OriginStamp before using the name in the package name?

- I'm really hesitant about the practice of trying to timestamp R objects rather than files.  Why does the method use `version=2` here instead of the default?  `serialize` is not suitable for archival purposes, as noted in the documentation, and having changed in the past and being subject to change in the future.  In UNIX philosophy, text files are the universal interface, and it seems that users would be better to serialize files rather than objects? I think when `storrr` does this, it first trims off the R version bit of the serialization to be more stable?  Anyway, just a thought, happy to be convinced otherwise!

- I think the vignette would be more compelling if it provided more narrative about the author's envisioned use cases and corresponding workflow.  For example, I was just discussing the issue of verifying timestamps with colleagues who are all interested in scientific and statistical methods for  _forecasting_.  They noted it would be desirable to verify that a forecast was produced at a particular time prior to when a measurement was actually made.  I understand that it often makes sense to defer to OriginStamp's own material about the many possible use cases, limitations, etc, but at least a little more information could be provided here.

- Likewise, I think it the vignette would be greatly helped by speaking to the issue of longevity, an area where I found the documentation from OriginStamp to be somewhat unclear.  At first glance, it is easy to assume that this ability to confirm a timestamp is only as good as the lifetime of the company behind the API.  In fact, it seems the designers have gone out of their way to make sure they are publishing timestamps into major commercial block chains, not merely maintaining there own.  As such, it seems like it should be possible to verify a timestamp even if OriginStamp is defunct.  But in practice, it seems like this would be very difficult to do, as all package functionality requires an OriginStamp API key.  Am I missing something?  It seems like in order to make it worthwhile to go through all this trouble to register timestamps in major digital currency blockchains, that it should be possible to verify those timestamps directly from the public block chains without an API key or third party of any kind (even if it is not feasible to submit new timestamps without a key).  If so, would it be possible to support that? If not, it kind of feels like "what's the point?" e.g. why not just rely on a trusted authority to gpg sign the hash or something?

- Lastly, it might be interesting to have some discussion of alternative approaches (or why things that seem like alternatives aren't).  Maybe this is too far beyond the scope of a vignette, but I'm just curious.
