### 2021-03-23 a205596
Basic structure for verification
