+++ man/transaction.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/extract_transaction.R
+\name{transaction}
+\alias{transaction}
+\title{Return the transaction id from the pdf certificate}
+\usage{
+transaction(x)
+}
+\arguments{
+\item{x}{filename or URL pointing either to the the pdf certificate issued from OriginStamp.}
+}
+\value{
+the transaction id
+}
+\description{
+Return the transaction id from the pdf certificate
+}
+\examples{
+transaction("https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/certificate.Bitcoin.pdf")
+
+}
