### 2021-03-24 73c5b6d
v_8.1 - added verification
Implementation of verification including basic vignette and tests
