### 2021-03-25 0cebd6d
Add verification of proof
