+++ R/extract_proof.R
@@ -1,8 +1,9 @@
 #' Extract the proof from an xml proof or pdf certificate
 #'
-#' If the v arification fails, the function aborts and raises an error.
+#' If the v verification fails, the function aborts and raises an error.
 #' @param x filename or URL pointing either to the xml proof or the pdf certificate issued from OriginStamp.
 #' @param verify if `FALSE`, read the root hash from the xml file. If `TRUE`, verify the root hash using the proof.
+#' @param verbose if `TRUE`, details on the verification will be printed which can help to see where the proof is wrong.
 #'
 #' @return the root hash
 #'
@@ -13,12 +14,16 @@
 #' @export
 #'
 #' @examples
-#' extract_proof("https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/proof.Bitcoin.xml")
-#' extract_proof("https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/certificate.Bitcoin.pdf")
+#' extract_proof(system.file("certificate.Bitcoin.pdf", package = "ROriginStamp"))
+#' extract_proof(system.file("proof.Bitcoin.xml", package = "ROriginStamp"))
+#' extract_proof(system.file("proof.Bitcoin.xml", package = "ROriginStamp"), cverify = FALSE)
+#' extract_proof(system.file("proof.faulty.xml", package = "ROriginStamp"))
+#' extract_proof(system.file("proof.faulty.xml", package = "ROriginStamp"), verbose = TRUE)
 #'
 extract_proof <- function(
   x,
-  verify = FALSE
+  verify = TRUE,
+  verbose = FALSE
 ) {
   type <- strsplit(x, "\\.")
   type <- tail(type[[1]], 1)
@@ -36,12 +41,14 @@ extract_proof <- function(
   if (!verify) {
     warning("\n#############################\n This proof is read from the document as provided.\n It is NOT verified for consistency!\n#############################")
   } else {
-    if (!verify_proof( xml )) {
+    if (!verify_proof( xml, verbose )) {
       stop(
         "\n",
-        "ATTENTION ATTENTION ATTENTION ATTENTION ATTENTION ATTENTION\n",
-        "   This is not a valid proof!!!\n",
-        "ATTENTION ATTENTION ATTENTION ATTENTION ATTENTION ATTENTION\n",
+        "#################################################################\n",
+        "## ATTENTION ATTENTION ATTENTION ATTENTION ATTENTION ATTENTION ##\n",
+        "##              This is not a valid proof!!!                   ##\n",
+        "## ATTENTION ATTENTION ATTENTION ATTENTION ATTENTION ATTENTION ##\n",
+        "#################################################################\n",
         "\n"
       )
     }
