### 2021-03-26 3e7824d
Add funder and regenerate pkgdown
