### 2021-05-11 #433 @cgries
- [x] **Examples** (that run successfully locally) for all exported functions

Estimated hours spent reviewing: 3h

---

I agree with everything @cboettig already commented on:

- I am not able to parse the messages returned by the functions.
- The vignette only discusses basic functionality and does not provide any real-world use cases; hence, I am a bit confused on how I would incorporate this into a workflow.
- The documentation and examples are minimal, and some are not very useful. E.g. `api_url()` Value: 'Either the old api url (when url is supplied) or the current api url (when url is not supplied)' did not help me to understand what URL I was supposed to provide.
- What is the expected time frame of a valid return from this third-party provider OriginStamp? Are we talking weeks, months, years for a user to get a verification certificate for a certain file?
- I agree that a discussion on why using this approach is preferable over any other way to assure authorship and timestamp of creation.

`devtools::check()` returns a warning:
checking for unstated dependencies in examples ... OK
   WARNING
  'qpdf' is needed for checks on size reduction of PDFs

`devtools::test()` mostly fails, I assume, because of local installation path and not accessing my api key?
