### 2021-05-17 #433 @rkrug
@cboettig Just in addition to the standard blockchain, OriginStamp also offers a [Media Blockchain](https://docs.originstamp.com/guide/#blockchain-currencies):

[What is a media blockchain](https://docs.originstamp.com/faq/#what-is-a-media-blockchain)?
Newspapers are distributed in the physical world like blockchains are in the digital. OriginStamp publishes the root hash of the Merkle Tree in different newspapers so that all potential newspaper readers become witnesses of the timestamp. Moreover, the root hash of today is a leaf hash for tomorrow, which means that the two Merkle Trees are linked and subsequent newspaper editions form a chain. The security and reliability of a media blockchain relies solely on the number of archived copies of each edition.

Intriguing.
