### 2021-05-25 aa3ce18
hash.file(): move close into on.exit()
