### 2021-06-08 0c109e3
Remove hashing of R objects (exception: character vectors)
