+++ tests/testthat/test_006_hash.R
@@ -2,20 +2,19 @@ test_that(
   "hash",
   {
     expect_snapshot_error(
-      as.hash(letters),
+      hash(NULL),
       cran = FALSE
     )
     expect_snapshot(
-      x = {
+      {
         tf <- tempfile()
         write.csv(data.frame(letters, LETTERS, 1:26), tf)
         x <- list(
           hash         = hash(as.hash("2c5d36be542f8f0e7345d77753a5d7ea61a443ba6a9a86bb060332ad56dba38e")),
           r_object     = hash(LETTERS),
-          r_object_def = hash.default(LETTERS),
-          file         = suppressMessages( hash(tf) ),
-          file_file    = suppressMessages( hash.file(tf) ),
-          null         = hash(NULL)
+          file         = hash(tf),
+          file_file    = hash.file(tf),
+          LETTERS      = as.hash(LETTERS)
         )
         unlink(tf)
         x
