### 2022-03-31 b277942
Removing of functions
