+++ man/txid_from_root_hash.Rd
@@ -1,24 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/txid_from_root_hash.R
-\name{txid_from_root_hash}
-\alias{txid_from_root_hash}
-\title{Get the transaction id from the root hash}
-\usage{
-txid_from_root_hash(root_hash)
-}
-\arguments{
-\item{root_hash}{root hash which has been submitted as OP_RETURN}
-}
-\value{
-a vector containing the transaction ids found. Usially only one
-}
-\description{
-This opnly results in valid returns for timestamps after mif February 2021.
-These are storing the root hash in OP_RETURN and these can therefore be searched.
-}
-\examples{
-txid_from_root_hash("bf6eec9f213846a53ed187d41d2b9ed2e954649a45d2883e5df6b107b2eb0b75")
-txid_from_root_hash("bf6eec9f213846a53ed187d41d2b9ed2e954649a45d2883e5dXXXXXXXXXXXXXX")
-txid_from_root_hash("love")
-
-}
