### 2022-08-09 #433 @noamross
@rkrug.  We're going back and picking up some reviews that got lost in ambiguous situations like this.  My apologies that this issue has foundered so long!  If you are still interested, I think the first solution is fine.  I will ask reviewers are amenable to reviewing your changes, and if not, we'll do an editors' review to check that you have addressed them.
