### 2022-08-10 c9574e5
Revert "Removing of functions"
This reverts commit b277942f0570737cb127e6ff73766ead47b67d98.
