+++ man/transaction_from_txid.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/transaction_from_txid.R
+\name{transaction_from_txid}
+\alias{transaction_from_txid}
+\title{Find transaction based on transaction ID}
+\usage{
+transaction_from_txid(txid)
+}
+\arguments{
+\item{txid}{a vector containing the transaction ids. Usiully only one}
+}
+\value{
+a list of the transactions identified by the txid
+}
+\description{
+Find transaction based on transaction ID
+}
+\examples{
+
+}
