+++ NAMESPACE
@@ -15,6 +15,9 @@ export(hash)
 export(hash.character)
 export(hash.file)
 export(new_OriginStampResponse)
+export(transaction_from_txid)
+export(txid_from_root_hash)
+export(verify_timestamp)
 importFrom(curl,curl_download)
 importFrom(curl,curl_fetch_memory)
 importFrom(curl,handle_setheaders)
