### 2022-08-10 21a1b90
Give error messages in not working functions which rely on disappeared host.
