### 2022-08-10 cd73f07
rename hash function and make vignette compile.
