+++ vignettes/usecases.Rmd
@@ -1,19 +1,0 @@
----
-title: "Usecases"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Usecases}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-```{r setup}
-library(ROriginStamp)
-```
