### 2022-08-10 884baf4
Fix errors in example and warnings and notes in check
