+++ man/extract_hash.Rd
@@ -16,7 +16,9 @@ an \code{hash} object containing the hash as used in the proof
 Return the hash of an xml proof or pdf certificate
 }
 \examples{
-extract_hash("https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/proof.Bitcoin.xml")
-extract_hash("https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/certificate.Bitcoin.pdf")
+fn <- "https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/proof.Bitcoin.xml"
+extract_hash(fn)
+fn <- "https://raw.githubusercontent.com/rkrug/ROriginStamp/master/inst/certificate.Bitcoin.pdf"
+extract_hash(fn)
 
 }
