+++ man/extract_proof.Rd
@@ -22,8 +22,11 @@ If the v verification fails, the function aborts and raises an error.
 \examples{
 extract_proof(system.file("certificate.Bitcoin.pdf", package = "ROriginStamp"))
 extract_proof(system.file("proof.Bitcoin.xml", package = "ROriginStamp"))
-extract_proof(system.file("proof.Bitcoin.xml", package = "ROriginStamp"), cverify = FALSE)
-extract_proof(system.file("proof.faulty.xml", package = "ROriginStamp"))
-extract_proof(system.file("proof.faulty.xml", package = "ROriginStamp"), verbose = TRUE)
+extract_proof(system.file("proof.Bitcoin.xml", package = "ROriginStamp"), verify = FALSE)
+######
+## These will produce errors when executed
+  # extract_proof(system.file("proof.faulty.xml", package = "ROriginStamp"))
+  # extract_proof(system.file("proof.faulty.xml", package = "ROriginStamp"), verbose = TRUE)
+######
 
 }
