### 2022-08-10 #433 @rkrug
I just merged my changes in the master branch and made a release v_0.9.0.

As suggested by @noamross, I have prepared the package as a "Proof of Concept". Therefore, I did not address most issues concerning the documentation.

I added stop() statements in the functions which do not work anymore due to a retired website and stated this in the man pages as well. If other sites can be found (and I am sure they are) which provide the same service for the lookup, the functions can be easily re-activated.

I also removed the possibility to hash R objects (including characters) as these are, as pointed out by the reviewers, dependant on the serialisation used by R, which can change over time.

I hope that these changes are sufficient, considering the "Proof of Concept" nature of the package.

Thanks,

Rainer
